change nodeinfo mps to float

+++ Tangle.Net/Entity/NodeInfo.cs
@@ -33,7 +33,7 @@
     public string Bech32Hrp { get; set; }
 
     [JsonProperty("messagesPerSecond")]
-    public int MessagesPerSecond { get; set; }
+    public float MessagesPerSecond { get; set; }
 
     [JsonProperty("pruningIndex")]
     public int PruningIndex { get; set; }
